refactor: Card 이름 변경
