+++ src/views/workspace/components/CardItem.vue
